restruct file tree
