+++ MAST/.gitignore
@@ -1,0 +1,3 @@
+Multivariate_ts/
+log/
+cricket_logs_*/
