+++ MAST/utils_MAST.py
@@ -163,7 +163,7 @@ def fill_out_with_Nan(data,max_length):
 
 def get_label_dict(file_path):
     label_dict ={}
-    with open(file_path) as file:
+    with open(file_path,encoding='utf-8') as file:
         lines = file.readlines()
         for line in lines:
             if '@classLabel' in line:
@@ -176,7 +176,7 @@ def get_label_dict(file_path):
 
 
 def get_data_and_label_from_ts_file(file_path,label_dict):
-    with open(file_path) as file:
+    with open(file_path,encoding='utf-8') as file:
         lines = file.readlines()
         Start_reading_data = False
         Label_list = []
